requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albumentations==1.1.0
altair==5.5.0
antlr4-python3-runtime==4.8
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
base58==2.1.1
bezier==2023.7.28
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
# Editable install with no version control (clip==1.0)
-e /data/shuangjun.du/work/FaceSwap/REFace/CLIP
cmake==4.1.0
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.4
decorator==4.4.2
diffusers==0.30.3
Dift==0.0.7
dlib @ file:///home/conda/feedstock_root/build_artifacts/dlib-split_1726093008991/work
docker-pycreds==0.4.0
easydict==1.13
einops==0.4.1
face-alignment==1.4.1
filelock==3.19.1
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
ftfy==6.0.3
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
glfw==2.7.0
grpcio==1.74.0
hf-xet==1.1.10
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
imutils==0.5.4
insightface==0.7.3
invisible-watermark==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kornia==0.6.0
lazy_loader==0.4
llvmlite==0.44.0
Markdown==3.9
MarkupSafe==3.0.2
matplotlib==3.10.7
ml_dtypes==0.5.3
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.6.4
narwhals==2.5.0
natsort==8.4.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
omegaconf==2.1.1
onnx==1.19.1
onnxruntime-gpu==1.16.3
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
packaging==24.1
pandas==2.0.3
pillow==11.3.0
platformdirs==4.4.0
prettytable==3.16.0
proglog==0.1.10
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pudb==2019.2
pydeck==0.9.1
pyDeprecate==0.3.1
Pygments==2.19.2
PyOpenGL==3.1.7
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytorch-lightning==1.4.2
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
qudida==0.0.4
# Editable install with no version control (REFace==0.0.1)
-e /data/shuangjun.du/work/FaceSwap/REFace
referencing==0.36.2
regex==2024.9.11
requests==2.32.3
rpds-py==0.27.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
sentry-sdk==2.37.1
setproctitle==1.3.7
six==1.17.0
smmap==5.0.2
streamlit==0.73.1
sympy==1.14.0
# Editable install with no version control (taming-transformers==0.0.1)
-e /data/shuangjun.du/work/FaceSwap/REFace/taming-transformers
tensorboard==2.20.0
tensorboard-data-server==0.7.2
test-tube==0.7.5
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.12.1
toml==0.10.2
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchmetrics==0.6.0
torchvision==0.14.1+cu117
tornado==6.5.2
tqdm==4.66.5
transformers==4.19.2
trimesh==4.4.9
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
urwid==3.0.2
validators==0.35.0
wandb==0.18.1
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0

